interface_aeropendulo/pyproject.toml
[tool.poetry]
name = "interface"
version = "0.1.0"
description = "Interface para interagit com o Aeropêndulo"
authors = ["Oseiasdfarias <oseias.dfarias@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
pyserial = "^3.5"
customtkinter = "^5.1.2"
pyudev = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
